requirements/base.txt
celery==5.2.3
boto3==1.18.16
Django==4.0.7
# django-admin-list-filter-dropdown==1.0.3
django-allauth==0.48.0
#django-allauth>=0.54.0
django-cors-headers==3.11.0
django-dbbackup==4.0.0b0
# django-crispy-forms==1.14.0
# crispy-bootstrap5==0.6
#django-dotenv==1.4.2
python-dotenv==1.0.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.7.1
django-leaflet==0.28.2
django-phonenumber-field[phonenumberslite]==6.1.0
#django-rest-auth==0.9.5
dj-rest-auth==2.2.3
django-storages==1.12.3
django-tables2==2.4.1
djangorestframework==3.13.1
djangorestframework-gis==0.18
djangorestframework-csv==2.1.1
# djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
numpy==1.22.2
pandas==1.4.1
psycopg2==2.9.3
